remove: per thread time printing

+++ code/include/util/benchmark.h
@@ -2,8 +2,6 @@
 #ifndef UTIL_BENCHMARK_H
 #define UTIL_BENCHMARK_H
 
-#include "system/types.h"
-
 void bench_io_start(void);
 void bench_align_start(void);
 void bench_filter_start(void);
@@ -16,6 +14,6 @@ void bench_io_print(void);
 void bench_align_print(void);
 void bench_filter_print(void);
 
-void bench_total_print(s64 alignments);
+void bench_total_print(double alignments);
 
 #endif /* UTIL_BENCHMARK_H */
